+++ yash-env/src/system/shared.rs
@@ -45,7 +45,6 @@ use crate::job::ProcessState;
 #[cfg(doc)]
 use crate::Env;
 use enumset::EnumSet;
-use nix::sys::stat::FileStat;
 use nix::sys::time::TimeSpec;
 use std::cell::RefCell;
 use std::convert::Infallible;
@@ -305,7 +304,7 @@ impl System for &SharedSystem {
     fn fstat(&self, fd: Fd) -> Result<Stat> {
         self.0.borrow().fstat(fd)
     }
-    fn fstatat(&self, dir_fd: Fd, path: &CStr, follow_symlinks: bool) -> Result<FileStat> {
+    fn fstatat(&self, dir_fd: Fd, path: &CStr, follow_symlinks: bool) -> Result<Stat> {
         self.0.borrow().fstatat(dir_fd, path, follow_symlinks)
     }
     fn is_executable_file(&self, path: &CStr) -> bool {
@@ -487,7 +486,7 @@ impl System for SharedSystem {
         (&self).fstat(fd)
     }
     #[inline]
-    fn fstatat(&self, dir_fd: Fd, path: &CStr, follow_symlinks: bool) -> Result<FileStat> {
+    fn fstatat(&self, dir_fd: Fd, path: &CStr, follow_symlinks: bool) -> Result<Stat> {
         (&self).fstatat(dir_fd, path, follow_symlinks)
     }
     #[inline]